Rename serializers/{base,mixins}.py
